% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rqdist.R
\name{augment.rqdist}
\alias{augment.rqdist}
\title{Augment Quantile Regression Distribution Predictions}
\usage{
augment.rqdist(object, newdata, rearrange = TRUE)
}
\description{
Augment Quantile Regression Distribution Predictions
}
