% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stepfun_helpers.R
\name{qf_stepfun_to_dst}
\alias{qf_stepfun_to_dst}
\title{Convert a Step Quantile Function to a Distribution}
\usage{
qf_stepfun_to_dst(qf_stepfun)
}
\arguments{
\item{qf_stepfun}{Step function representing a quantile function.}
}
\value{
Discrete distionary/distplyr distribution.
If the step function provided has a portion that
decreases, NULL is returned.
}
\description{
Convert a Step Quantile Function to a Distribution
}
