% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rqdist.R
\name{rqdist}
\alias{rqdist}
\title{Linear Quantile Regression Process Distribution Model}
\usage{
rqdist(formula, data, ...)
}
\arguments{
\item{formula}{An object of class "formula".}

\item{data}{Data frame containing the data.}

\item{...}{Other parameters to pass to \code{quantreg::rq()},
aside from the \code{tau} option.}
}
\value{
An object of class "rqdist" that returns
distributions of class "dst" from the distplyr package.
}
\description{
Fits a quantile regression process model based on an
assumption of linear conditional quantiles. A wrapper
for \code{quantreg::rq()}.
}
