% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rqdist.R
\name{stepqf_to_dst}
\alias{stepqf_to_dst}
\title{Convert a Step Quantile Function to a Distribution}
\usage{
stepqf_to_dst(stepqf, ...)
}
\arguments{
\item{stepqf}{Step function representing a quantile function.}
}
\value{
Object of class "stepdst" from the \code{distplyr}
package. If the step function provided has a portion that
decreases, a distribution object ("dst") is still output,
but only the quantile function is available.
}
\description{
Convert a Step Quantile Function to a Distribution
}
