% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rqdist.R
\name{stepqf_to_stepdst}
\alias{stepqf_to_stepdst}
\title{Convert a Step Quantile Function to a Step Distribution}
\usage{
stepqf_to_stepdst(stepqf, ...)
}
\arguments{
\item{stepqf}{Step function representing a quantile function.}

\item{...}{Other arguments to pass to the \code{stepdst} or
\code{dst} function (for example, a density).}
}
\value{
Object of class "stepdst" from the \code{distplyr}
package. If the step function provided has a portion that
decreases, a distribution object ("dst") is still output,
but only the quantile function is available.
}
\description{
Convert a Step Quantile Function to a Step Distribution
}
