% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stepfun_helpers.R
\name{swap_step_continuity_direction}
\alias{swap_step_continuity_direction}
\title{Swap continuity direction of a step function}
\usage{
swap_step_continuity_direction(stepfun)
}
\arguments{
\item{stepfun}{Object of class "stepfun".}
}
\value{
The same `stepfun` object with the continuity direction swapped
between left and right.
}
\description{
Swap continuity direction of a step function
}
