% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rqdist.R
\name{predict.rqdist}
\alias{predict.rqdist}
\alias{augment.rqdist}
\title{Predict Quantile Regression Distributions}
\usage{
\method{predict}{rqdist}(object, newdata, rearrange = TRUE, ...)

\method{augment}{rqdist}(object, newdata, rearrange = TRUE)
}
\arguments{
\item{object}{Object of class "rqdist", returned from the
\code{rqdist()} function.}

\item{newdata}{Data frame from which to make predictions. If empty,
uses the training data.}

\item{rearrange}{The linear assumption sometimes results in
distributions with non-monotonic quantile functions and cdf's.
If \code{TRUE} (the default), uses the \code{quantreg::rearrange()} function
to make these functions non-decreasing (and therefore a legitimate
distribution). If \code{FALSE}, the quantile function is left as-is, and
may not represent a legitimate distribution.}

\item{...}{Not used.}
}
\value{
List of distributions ("dst" objects) corresponding
to each row of \code{newdata}, corresponding to the
estimated distribution of the response given the covariates
in \code{newdata}.
}
\description{
Predict distributions from a Linear Quantile Regression
Process Model. A wrapper around
\code{quantreg::predict.rq.process()},
bundling the distribution functions into a "dst" object.
}
